feat(compiler): error for duplicate enum
Added an error if an enum is defined twice. This is just a simple version of the check - once we've
got proper symbol resolution in place (#20) we'll want to also check for the situation where
the same enum is defined in separate files where the C# namespace is the same.

Issue: #9