chore: don't run tests on commit
A couple of the tests I'm about to add involve running some C# code as a script. Those two tests
alone double the time it's taking to run the tests on my machine from around 3 to 6 seconds,
which is getting into the realms of too long to run on each commit.