chore: add test explorer settings
Updating the settings.json file to add some settings for the .NET Test Explorer extension. It feels
a bit weird to add config settings for an extension that people might not be using, but I think the
settings will just be ignored if someone isn't using the extension, so I doubt it's a problem.