refactor(compiler): move type resolution to symbol
I realised that the way I'd designed things wasn't going to work, so I've simplified the binders so
that they just deal with creating the symbol for the immediate node, and child nodes are implemented
with lazy loaded properties. I've also moved type resolution to the symbols so that they can walk up
the symbol tree to find the correct type.