refactor(compiler): fix CollectionType inheritance
Lists, sets and maps all share a lot of common properties - the main difference is that lists and
sets have a single element type, whereas maps have a key type and a value type. Because of this
I've pulled the list and set specific code into a class called `ListOrSetType` so that all three
could share `CollectionType` and avoid duplicating all the shared code.

In addition:

- Changed the implementation of `CollectionType.Name` to just use the parsed text from Antlr. This
  means we lose any whitespace, but I don't think that's a problem for now.
- Fixed a couple of mistakes in the list and set collection tests where they had hard-coded input.

Part of #68