ci: fix source branch sha for commitlint
The approach I took to get the target branch doesn't work when the PR is coming from a clone.
The reason it was using `git rev-parse` is because for PRs, commitlint would end up failing
because it would check the merge commit automatically created by GitHub.

I've changed it to get the commit SHA immediately before the merge commit as the source
commit.

Fixes #74