fix(docs): fix test and coverage badge links
The unit test and code coverage badges were just linking to the badge image. I've altered them to
link to the build results instead. I tried to figure out how to link directly to the test or
coverage tabs, but it wasn't obvious so I've just left it linking to the main pipeline results page.

Part of #83